rename. HomeIcon.tsx , global.d.ts => global.d.ts, siwper.d.ts로 구분
